requirements.txt
django
djangorestframework
pillow
django-jsonfield
django-haystack
drf-haystack
whoosh
